# falconapi/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "FalconApi",
        "title": "Сервис админки FalconApi",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.100.155:3006",
    "basePath": "/api/v1",
    "paths": {
        "/generate-otp": {
            "post": {
                "description": "Генерация OTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Метод генарации OTP",
                "parameters": [
                    {
                        "description": "Generate OTP data",
                        "name": "GenerateOtp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usermgmtuc.GenerateOtpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PNG image data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Вход пользователя под логином и паролем",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Метод входа пользователя",
                "parameters": [
                    {
                        "description": "Login data",
                        "name": "LoginInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usermgmtuc.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usermgmtuc.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/region": {
            "get": {
                "description": "Получение списка регионов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Terminals"
                ],
                "summary": "Метод получения списка регионов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/entities.TRegion"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/terminalsinfo": {
            "get": {
                "description": "Получение инфо - статусов терминалов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Terminals"
                ],
                "summary": "Метод получения инфо - статусов терминалов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/entities.TerminalStatus"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/terminalstatuses": {
            "get": {
                "description": "Получение статусов терминалов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Terminals"
                ],
                "summary": "Метод получения статусов терминалов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/entities.TerminalStatus"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "Регистрация пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Метод регистрации пользователя",
                "parameters": [
                    {
                        "description": "Login data",
                        "name": "RegisterInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usermgmtuc.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/validate-otp": {
            "post": {
                "description": "Валидация OTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Метод валидации OTP",
                "parameters": [
                    {
                        "description": "Validate OTP data",
                        "name": "ValidateOtp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usermgmtuc.ValidateOtpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usermgmtuc.ValidateOtpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.TRegion": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entities.TerminalStatus": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "detailStatus": {
                    "type": "string"
                },
                "endpointDisabled": {
                    "type": "boolean"
                },
                "endpointID": {
                    "type": "integer"
                },
                "endpointNum": {
                    "type": "integer"
                },
                "lastPayment": {
                    "type": "string"
                },
                "lastPaymentDetail": {
                    "type": "string"
                },
                "lastPing": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "regionName": {
                    "type": "string"
                },
                "region_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "usermgmtuc.GenerateOtpRequest": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
        "usermgmtuc.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
        "usermgmtuc.LoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "otpGenerated": {
                    "type": "boolean"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "usermgmtuc.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                },
                "lastName": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                },
                "mobileNumber": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
        "usermgmtuc.ValidateOtpRequest": {
            "type": "object",
            "properties": {
                "otp_token": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "usermgmtuc.ValidateOtpResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "QrCode       bytes.Buffer `json:\"qr_code,omitempty\"`",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}